refactor(tenant-members): reorganize factory imports and remove deprecated files
- Updated import paths for `TenantMemberAggregateFactory` and `TenantMemberViewModelFactory` to reflect new directory structure.
- Removed obsolete factory files for tenant member aggregate and view model, streamlining the codebase.
- Ensured all references to tenant member view models are consistent with the new structure.
