+++ apps/api/src/tenant-context/tenant-members/application/queries/tenant-member-find-by-criteria/tenant-member-find-by-criteria.query-handler.ts
@@ -3,7 +3,7 @@ import {
   TENANT_MEMBER_READ_REPOSITORY_TOKEN,
   TenantMemberReadRepository,
 } from '@/tenant-context/tenant-members/domain/repositories/tenant-member-read.repository';
-import { TenantMemberViewModel } from '@/tenant-context/tenant-members/domain/view-models/tenant-member.view-model';
+import { TenantMemberViewModel } from '@/tenant-context/tenant-members/domain/view-models/tenant-member/tenant-member.view-model';
 import { Inject, Logger } from '@nestjs/common';
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { FindTenantMembersByCriteriaQuery } from './tenant-member-find-by-criteria.query';
