+++ apps/api/src/tenant-context/tenant-members/domain/factories/tenant-member-view-model/tenant-member-view-model.factory.ts
@@ -2,7 +2,7 @@ import { IReadFactory } from '@/shared/domain/interfaces/read-factory.interface'
 import { TenantMemberAggregate } from '@/tenant-context/tenant-members/domain/aggregates/tenant-member.aggregate';
 import { ITenantMemberCreateViewModelDto } from '@/tenant-context/tenant-members/domain/dtos/view-models/tenant-member-create/tenant-member-create-view-model.dto';
 import { TenantMemberPrimitives } from '@/tenant-context/tenant-members/domain/primitives/tenant-member.primitives';
-import { TenantMemberViewModel } from '@/tenant-context/tenant-members/domain/view-models/tenant-member.view-model';
+import { TenantMemberViewModel } from '@/tenant-context/tenant-members/domain/view-models/tenant-member/tenant-member.view-model';
 import { Injectable, Logger } from '@nestjs/common';
 
 /**
