+++ apps/api/src/tenant-context/tenant-members/domain/repositories/tenant-member-read.repository.ts
@@ -1,6 +1,6 @@
 import { Criteria } from '@/shared/domain/entities/criteria';
 import { PaginatedResult } from '@/shared/domain/entities/paginated-result.entity';
-import { TenantMemberViewModel } from '@/tenant-context/tenant-members/domain/view-models/tenant-member.view-model';
+import { TenantMemberViewModel } from '@/tenant-context/tenant-members/domain/view-models/tenant-member/tenant-member.view-model';
 
 export const TENANT_MEMBER_READ_REPOSITORY_TOKEN = Symbol(
   'TenantMemberReadRepository',
