+++ apps/api/src/tenant-context/tenant-members/infrastructure/database/mongodb/repositories/tenant-member-mongodb.repository.ts
@@ -3,7 +3,7 @@ import { PaginatedResult } from '@/shared/domain/entities/paginated-result.entit
 import { BaseMongoRepository } from '@/shared/infrastructure/database/mongodb/base-mongo/base-mongo.repository';
 import { MongoService } from '@/shared/infrastructure/database/mongodb/services/mongo.service';
 import { TenantMemberReadRepository } from '@/tenant-context/tenant-members/domain/repositories/tenant-member-read.repository';
-import { TenantMemberViewModel } from '@/tenant-context/tenant-members/domain/view-models/tenant-member.view-model';
+import { TenantMemberViewModel } from '@/tenant-context/tenant-members/domain/view-models/tenant-member/tenant-member.view-model';
 import { TenantMemberMongoDBMapper } from '@/tenant-context/tenant-members/infrastructure/database/mongodb/mappers/tenant-member-mongodb.mapper';
 import { Injectable, Logger } from '@nestjs/common';
