+++ apps/api/src/tenant-context/tenant-members/infrastructure/database/prisma/mappers/tenant-member-prisma.mapper.ts
@@ -1,5 +1,5 @@
 import { TenantMemberAggregate } from '@/tenant-context/tenant-members/domain/aggregates/tenant-member.aggregate';
-import { TenantMemberAggregateFactory } from '@/tenant-context/tenant-members/domain/factories/tenant-member-aggregate.factory';
+import { TenantMemberAggregateFactory } from '@/tenant-context/tenant-members/domain/factories/tenant-member-aggregate/tenant-member-aggregate.factory';
 import { TenantMemberPrismaDto } from '@/tenant-context/tenant-members/infrastructure/database/prisma/dtos/tenant-member-prisma.dto';
 import { Injectable, Logger } from '@nestjs/common';
 import { TenantMemberRoleEnum } from '@prisma/client';
