+++ apps/api/src/tenant-context/tenant-members/transport/graphql/mappers/tenant-member.mapper.ts
@@ -1,5 +1,5 @@
 import { PaginatedResult } from '@/shared/domain/entities/paginated-result.entity';
-import { TenantMemberViewModel } from '@/tenant-context/tenant-members/domain/view-models/tenant-member.view-model';
+import { TenantMemberViewModel } from '@/tenant-context/tenant-members/domain/view-models/tenant-member/tenant-member.view-model';
 import {
   PaginatedTenantMemberResultDto,
   TenantMemberResponseDto,
